requirements.txt
dominate
torchgeometry
func-timeout
tqdm
matplotlib
opencv-python
lmdb
numpy
GPUtil
pillow
scikit-learn==0.23.2
visdom
ninja

scipy

scikit-image
lpips
natsort
openmesh
pyqt5
